Class code cleanup - removed obsolete components.
Moved profile mgmt for class into new class-form component.
