--- client/src/app/class/class-forms-player/class-forms-player.component.css
@@ -1,10 +1,0 @@
-#profile-mgmt {
-  position: relative;
-  top: -100px;
-  left: 50px;
-}
-#profile-toolbar {
-  display: flex;
-  position: absolute;
-  top: -80px;
-}
