--- client/src/app/class/class-forms-player/class-forms-player.component.html
@@ -1,16 +1,0 @@
-<div #container> </div>
-<div id="profile-mgmt" *ngIf="curriculum === 'student-registration' && studentId">
-  <div id="profile-toolbar">
-    <div (click)="$event ? archiveStudent(studentId) : null" >
-      <button routerLink="dashboard" mat-menu-item>
-      <mat-icon class="material-icons menu-tangy-location-list-icon">delete</mat-icon>
-    </button>
-    </div>
-    <div (click)="$event ? enableEditing(studentId, classId, $event) : null" >
-    <button mat-menu-item>
-      <mat-icon class="material-icons menu-tangy-location-list-icon">edit</mat-icon>
-    </button>
-    </div>
-  </div>
-</div>
-
